2025/03/04 refactor - finish

+++ src/test/java/com/example/assignment/repository/BTCPriceHistoryRepositoryTest.java
@@ -58,7 +58,7 @@ class BTCPriceHistoryRepositoryTest {
     Optional<BTCPriceHistory> latestRecord =
         btcPriceHistoryRepository.findTopByOrderByTimestampDesc();
     assertThat(latestRecord).isPresent();
-    assertThat(latestRecord.get().getPrice()).isEqualTo(450.0);
+    assertThat(latestRecord.get().getPrice()).isEqualTo(BigDecimal.valueOf(450.0));
     assertThat(latestRecord.get().getTimestamp()).isEqualTo(savedRecord.getTimestamp());
   }
